prijs wordt aangepast als helmet true is. daarnaast spdPedals toegevoegd aan de booking. ook is het nu mogelijk om 1 booking op te vragen dmv id booking/{id}

+++ src/main/java/nl/mtbrental/eindproject/dto/BikeDto.java
@@ -5,7 +5,7 @@ import nl.mtbrental.eindproject.model.Bike;
 public class BikeDto {
     public Long id;
     public String bikeName;
-    public Long quantityTotal;
+    public int quantityTotal;
     public Long pricePerDay;
 
     public static BikeDto fromBike(Bike bike) {
