+++ src/main/java/nl/mtbrental/eindproject/dto/BikeInputDto.java
@@ -4,7 +4,7 @@ import nl.mtbrental.eindproject.model.Bike;
 
 public class BikeInputDto {
     public String bikeName;
-    public Long quantityTotal;
+    public int quantityTotal;
     public Long pricePerDay;
 
     public Bike toBike() {
